requirement.txt
Django==1.11.6
mysqlclient==1.3.12
django-crispy-forms==1.7.0
django-import-export==0.5.1
django-reversion==2.0.0
django-formtools==2.1
future==0.16.0
httplib2==0.9.2
six==1.11.0
Pillow==4.3.0
djangorestframework==3.7.1
django-imagekit==4.0.1
